+++ Commands/basicCommands.py
@@ -1,5 +1,9 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+# A library that provides functionality to the @CoreDumped_EventsBot
+# Copyright (C) 2017-2018
+# Javier Gines Sanchez <software@javisite.com>
+#
 
 import logging												## System module
 log = logging.getLogger(__name__)
@@ -18,7 +22,7 @@ import Functions.message as ms								## Own module
 def start(bot, update):
 	bd.startWithCommand(bot, update)
 
-	bot.sendMessage(chat_id=bd.chat_id, text=ms.helpOrStart, reply_to_message_id=bd.message.message_id)
+	bot.sendMessage(chat_id=bd.chat_id, text=ms.helpOrStart, reply_to_message_id=bd.message.message_id, disable_web_page_preview=True, parse_mode="MARKDOWN")
 
 
 # Command /restartP or /rebootP (Private)
@@ -60,17 +64,20 @@ def stopP(bot, update):
 def leaveGroup(bot, update):
 	bd.startWithCommand(bot, update)
 
-	if bd.user_id == bd.chatIDDeveloper:
-		if update.effective_chat != None and update.effective_chat.type != "private":
+	if update.effective_chat != None and update.effective_chat.type != "private":
+		admin = False
+		for adminMember in bd.message.chat.get_administrators():
+			if adminMember['user']['id'] is bd.user_id:
+				admin = True
+
+		if bd.user_id == bd.chatIDDeveloper or admin or bd.message.chat.all_members_are_administrators:
 			bot.sendMessage(chat_id=bd.chat_id, text=ms.leaving, reply_to_message_id=bd.message.message_id)
 			bot.getChat(chat_id=bd.chat_id).leave()
-
 		else:
-			bot.sendMessage(chat_id=bd.chat_id, text=ms.notGroupLeave, reply_to_message_id=bd.message.message_id)
-
+			bot.sendMessage(chat_id=bd.chat_id, text=ms.notAdmin[randint(0, len(ms.notAdmin)-1)], reply_to_message_id=bd.message.message_id)
+			bd.userNotAuthorizedMessage(bot, update)
 	else:
-		bot.sendMessage(chat_id=bd.chat_id, text=ms.notAdmin[randint(0, len(ms.notAdmin)-1)], reply_to_message_id=bd.message.message_id)
-		bd.userNotAuthorizedMessage(bot, update)
+		bot.sendMessage(chat_id=bd.chat_id, text=ms.notGroupLeave, reply_to_message_id=bd.message.message_id)
 
 
 # Changelog command /changelog