+++ Functions/basicData.py
@@ -1,5 +1,9 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+# A library that provides functionality to the @CoreDumped_EventsBot
+# Copyright (C) 2017-2018
+# Javier Gines Sanchez <software@javisite.com>
+#
 
 import logging												## System module
 log = logging.getLogger(__name__)
@@ -38,14 +42,13 @@ def startWithCommand(bot, update, args=['']):
 	user_id = message.from_user.id
 
 	log.info(message.text.split(' ')[0] + ' ' + ' '.join(args) + ' --> ' + username + " (chat_id:" + str(chat_id) + " , user_id:"+ str(user_id) + ")")
-#	bot.sendMessage(chat_id=chatIDDeveloper, text=message.text.split(' ')[0] + ' ' + ' '.join(args) + ' --> ' + username + " (chat_id:" + str(chat_id) + " , user_id:"+ str(user_id) + ")")
 
 def userNotAuthorizedMessage(bot, update, args=['']):
 	bot.sendMessage(chat_id=chatIDDeveloper, text=ms.userNotAuthorizedCommand.replace("$args1",
 	message.text.split(' ')[0] + ' ' + ' '.join(args) + ' --> ' + username + " (chat_id:" + str(chat_id) + " , user_id:"+ str(user_id) + ")"))
 
 def exceptionHandler(bot, update, nameModule, exception, args=['']):
-	log.error(str(exception)+ " - Module: " + nameModule + " Line "+ str(exc_info()[2].tb_lineno))
+	log.error(str(exception)+ " - Module: " + nameModule + " Line "+ (str(exc_info()[2].tb_lineno) if exc_info()[2] != None else "None"))
 	bot.sendMessage(chat_id=chat_id, text=ms.errorExecCommandUser, reply_to_message_id=message.message_id)
 	bot.sendMessage(chat_id=chatIDDeveloper, text=ms.errorExecCommandAdmin.replace("$args1",
 	message.text.split(' ')[0] + ' ' + ' '.join(args) + ' --> ' + username + " (chat_id:" + str(chat_id) + " , user_id:"+ str(user_id) + ")"))